Extract PrivateKeyFormatter into a component
